pc: add USE_TPM define to cflags, if enabled
The USE_TPM #ifdef guards exist in the public XTT headers, so users of
the library must add this definition if the local XTT was compiled
with TPM support.

For anyone using CMake, xtt-config.cmake sets this correctly.

This commit extends the same support for pkg-config users.  The output
of $(pkg-config --cflags xtt) will include "-DUSE_XTT" if XTT was
compiled with TPM support.