util: Restructure (and fix!) x.509 and ASN.1 operations.
NOTE: This changes the public API.
Due to the fix below, the length of a new x.509 cert is variable, so the
resulting length needed to be returned to the user
(and the old `xtt_x509_certificate_length` function and matching #define
needed to be changed to `xtt_x509_certificate_max_length`).

This PR fixes the x.509 certificate generation, by storing the signature
correctly.

This also restructures the file-structure of the internal ASN.1 and
x.509 logic, and changes these operations from overly-complex functions
to hardcoded char-buffers.
This should be easier to understand (and thus less prone to error and
easier to alter if needed).