test: Replace hardcoded cert test with OpenSSL-based script.
The previous test assumed a lot about the implementation
of the x.509 cert creation (and thus no longer really works as-is),
and really wasn't testing the cert creation but rather details
of its implementation.

The new test uses the openssl CLI, to test that our certs are accepted
by third-parties.

To keep the tests running on platforms without OpenSSL,
this test is gated behind a CMake option.